study(step3): 난수 생성 로직 분리
