+++ src/main/kotlin/racingcar/RandomCountGenerator.kt
@@ -1,0 +1,7 @@
+package racingcar
+
+import kotlin.random.Random
+
+object RandomCountGenerator {
+    fun generateCount() = Random(System.currentTimeMillis()).nextInt(0, 10)
+}
