+++ src/main/kotlin/racingcar/model/CarRacer.kt
@@ -15,8 +15,6 @@ class CarRacer {
         path += CHARACTER_MOVE
     }
 
-    fun generateCount() = Random(System.currentTimeMillis()).nextInt(0, 10)
-
     companion object {
         const val CHARACTER_MOVE = "-"
     }
