+++ src/main/kotlin/racingcar/model/CarRacingGame.kt
@@ -1,7 +1,24 @@
 package racingcar.model
 
+import racingcar.RandomCountGenerator
+
 /**
  * 자동차 경주
  */
-class CarRacingGame {
+class CarRacingGame(
+    private val numberOfGame: Int,
+    private val list: List<CarRacer>
+) {
+
+    fun start() {
+        for (i in 0..numberOfGame) {
+            startGame()
+        }
+    }
+
+    private fun startGame() {
+        for (racer in list) {
+            racer.move(RandomCountGenerator.generateCount())
+        }
+    }
 }
